# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Rental Housing Law Navigator",
    "description": "Not legal advice. Coverage is not a finding of compliance or a violation.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/addresses": {
      "get": {
        "summary": "Addresses",
        "operationId": "addresses_api_v1_addresses_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 200,
              "default": "",
              "title": "Q"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 25,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddressPage"
                }
              }
            }
          },
          "404": {
            "description": "Unknown ID"
          },
          "422": {
            "description": "Invalid request"
          },
          "503": {
            "description": "Dataset or extracted rules unavailable"
          }
        }
      }
    },
    "/api/v1/lookup": {
      "post": {
        "summary": "Address Lookup",
        "operationId": "address_lookup_api_v1_lookup_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LookupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LookupResponse"
                }
              }
            }
          },
          "404": {
            "description": "Unknown ID"
          },
          "422": {
            "description": "Invalid request"
          },
          "503": {
            "description": "Dataset or extracted rules unavailable"
          }
        }
      }
    },
    "/api/v1/lookup/assist": {
      "post": {
        "summary": "Assisted Lookup",
        "operationId": "assisted_lookup_api_v1_lookup_assist_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssistRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssistResponse"
                }
              }
            }
          },
          "404": {
            "description": "Unknown ID"
          },
          "422": {
            "description": "Invalid request"
          },
          "503": {
            "description": "Dataset or extracted rules unavailable"
          },
          "502": {
            "description": "Core output violated the shared contract"
          }
        }
      }
    },
    "/api/v1/lookup/evidence-package": {
      "post": {
        "summary": "Evidence Package",
        "operationId": "evidence_package_api_v1_lookup_evidence_package_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvidencePackageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvidencePackage"
                }
              }
            }
          },
          "404": {
            "description": "Unknown ID"
          },
          "422": {
            "description": "Invalid request"
          },
          "503": {
            "description": "Dataset or extracted rules unavailable"
          },
          "502": {
            "description": "Core output violated the shared contract"
          }
        }
      }
    },
    "/api/v1/facts": {
      "get": {
        "summary": "Fact Definitions",
        "operationId": "fact_definitions_api_v1_facts_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "$ref": "#/components/schemas/FactDefinition"
                  },
                  "type": "object",
                  "title": "Response Fact Definitions Api V1 Facts Get"
                }
              }
            }
          },
          "404": {
            "description": "Unknown ID"
          },
          "422": {
            "description": "Invalid request"
          },
          "503": {
            "description": "Dataset or extracted rules unavailable"
          }
        }
      }
    },
    "/api/v1/rules/{rule_id}/evidence": {
      "get": {
        "summary": "Rule Evidence",
        "operationId": "rule_evidence_api_v1_rules__rule_id__evidence_get",
        "parameters": [
          {
            "name": "rule_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Rule Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvidenceReport"
                }
              }
            }
          },
          "404": {
            "description": "Unknown ID"
          },
          "422": {
            "description": "Invalid request"
          },
          "503": {
            "description": "Dataset or extracted rules unavailable"
          }
        }
      }
    },
    "/api/v1/sources/{doc_id}/context": {
      "get": {
        "summary": "Source Context",
        "operationId": "source_context_api_v1_sources__doc_id__context_get",
        "parameters": [
          {
            "name": "doc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Doc Id"
            }
          },
          {
            "name": "start",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Start"
            }
          },
          {
            "name": "end",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "End"
            }
          },
          {
            "name": "max_depth",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 4,
              "minimum": 0,
              "default": 2,
              "title": "Max Depth"
            }
          },
          {
            "name": "max_chars",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 48000,
              "minimum": 100,
              "default": 24000,
              "title": "Max Chars"
            }
          },
          {
            "name": "max_spans",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 24,
              "minimum": 1,
              "default": 12,
              "title": "Max Spans"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceContext"
                }
              }
            }
          },
          "404": {
            "description": "Unknown ID"
          },
          "422": {
            "description": "Invalid request"
          },
          "503": {
            "description": "Dataset or extracted rules unavailable"
          }
        }
      }
    },
    "/api/v1/rules/{rule_id}": {
      "get": {
        "summary": "Rule Detail",
        "operationId": "rule_detail_api_v1_rules__rule_id__get",
        "parameters": [
          {
            "name": "rule_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Rule Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleDetail"
                }
              }
            }
          },
          "404": {
            "description": "Unknown ID"
          },
          "422": {
            "description": "Invalid request"
          },
          "503": {
            "description": "Dataset or extracted rules unavailable"
          }
        }
      }
    },
    "/api/v1/sources/{doc_id}": {
      "get": {
        "summary": "Source Detail",
        "operationId": "source_detail_api_v1_sources__doc_id__get",
        "parameters": [
          {
            "name": "doc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Doc Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceDocument"
                }
              }
            }
          },
          "404": {
            "description": "Unknown ID"
          },
          "422": {
            "description": "Invalid request"
          },
          "503": {
            "description": "Dataset or extracted rules unavailable"
          }
        }
      }
    },
    "/api/v1/changes": {
      "post": {
        "summary": "Changes",
        "operationId": "changes_api_v1_changes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangeResult"
                }
              }
            }
          },
          "404": {
            "description": "Unknown ID"
          },
          "422": {
            "description": "Invalid request"
          },
          "503": {
            "description": "Dataset or extracted rules unavailable"
          }
        }
      }
    },
    "/api/v1/changes/summary": {
      "post": {
        "summary": "Summarized Changes",
        "operationId": "summarized_changes_api_v1_changes_summary_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangeSummary"
                }
              }
            }
          },
          "404": {
            "description": "Unknown ID"
          },
          "422": {
            "description": "Invalid request"
          },
          "503": {
            "description": "Dataset or extracted rules unavailable"
          }
        }
      }
    },
    "/api/v1/source-comparisons": {
      "get": {
        "summary": "Compared Sources",
        "operationId": "compared_sources_api_v1_source_comparisons_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceComparisonsResponse"
                }
              }
            }
          },
          "404": {
            "description": "Unknown ID"
          },
          "422": {
            "description": "Invalid request"
          },
          "503": {
            "description": "Dataset or extracted rules unavailable"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AddressItem": {
        "properties": {
          "property": {
            "$ref": "#/components/schemas/PropertyFacts"
          },
          "resolution": {
            "$ref": "#/components/schemas/JurisdictionResolution"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "property",
          "resolution"
        ],
        "title": "AddressItem"
      },
      "AddressPage": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/AddressItem"
            },
            "type": "array",
            "title": "Items"
          },
          "disclaimer": {
            "type": "string",
            "title": "Disclaimer"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "total",
          "offset",
          "limit",
          "items",
          "disclaimer"
        ],
        "title": "AddressPage"
      },
...
      "AssistRequest": {
        "properties": {
          "address_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address Id"
          },
          "address": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RawAddress"
              },
              {
                "type": "null"
              }
            ]
          },
          "as_of": {
            "type": "string",
            "format": "date",
            "title": "As Of",
            "default": "2026-10-01"
          },
          "supplemental_facts": {
            "additionalProperties": true,
            "type": "object",
            "title": "Supplemental Facts"
          },
          "answers": {
            "items": {
              "$ref": "#/components/schemas/SupplementalAnswer"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Answers"
          },
          "scenario_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Scenario Id"
          },
          "limits": {
            "$ref": "#/components/schemas/AnalysisLimits"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "AssistRequest"
      },
      "AssistResponse": {
        "properties": {
          "lookup": {
            "$ref": "#/components/schemas/LookupResponse"
          },
          "question_plan": {
            "$ref": "#/components/schemas/QuestionPlan"
          },
          "evidence_reports": {
            "items": {
              "$ref": "#/components/schemas/EvidenceReport"
            },
            "type": "array",
            "title": "Evidence Reports"
          },
          "encoded_rules": {
            "items": {
              "$ref": "#/components/schemas/EncodedRuleRendering"
            },
            "type": "array",
            "title": "Encoded Rules"
          },
          "answers_applied": {
            "items": {
              "$ref": "#/components/schemas/SupplementalAnswer"
            },
            "type": "array",
            "title": "Answers Applied"
          },
          "scenario_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scenario Id"
          },
          "capabilities": {
            "additionalProperties": {
              "type": "string",
              "enum": [
                "implemented",
                "dependency_unavailable"
              ]
            },
            "type": "object",
            "title": "Capabilities"
          },
          "mode": {
            "type": "string",
            "enum": [
              "dataset",
              "synthetic",
              "contract_fixture"
            ],
            "title": "Mode"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "lookup",
          "question_plan",
          "evidence_reports",
          "encoded_rules",
          "answers_applied",
          "capabilities",
          "mode"
        ],
        "title": "AssistResponse"
      },
...
      "LookupRequest": {
        "properties": {
          "address_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address Id"
          },
          "address": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RawAddress"
              },
              {
                "type": "null"
              }
            ]
          },
          "as_of": {
            "type": "string",
            "format": "date",
            "title": "As Of",
            "default": "2026-10-01"
          },
          "supplemental_facts": {
            "additionalProperties": true,
            "type": "object",
            "title": "Supplemental Facts"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "LookupRequest"
      },
      "LookupResponse": {
        "properties": {
          "address": {
            "$ref": "#/components/schemas/PropertyFacts"
          },
          "as_of": {
            "type": "string",
            "format": "date",
            "title": "As Of"
          },
          "jurisdiction": {
            "$ref": "#/components/schemas/JurisdictionResolution"
          },
          "evaluations": {
            "items": {
              "$ref": "#/components/schemas/Evaluation"
            },
            "type": "array",
            "title": "Evaluations"
          },
          "rules": {
            "items": {
              "$ref": "#/components/schemas/Rule"
            },
            "type": "array",
            "title": "Rules"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/SourceDocument"
            },
            "type": "array",
            "title": "Sources"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          },
          "disclaimer": {
            "type": "string",
            "title": "Disclaimer"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "address",
          "as_of",
          "jurisdiction",
          "evaluations",
          "rules",
          "sources",
          "warnings",
          "metadata",
          "disclaimer"
        ],
        "title": "LookupResponse"
      },
...
      "RuleDetail": {
        "properties": {
          "rule": {
            "$ref": "#/components/schemas/Rule"
          },
          "versions": {
            "items": {
              "$ref": "#/components/schemas/Rule"
            },
            "type": "array",
            "title": "Versions"
          },
          "disclaimer": {
            "type": "string",
            "title": "Disclaimer"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "rule",
          "versions",
          "disclaimer"
        ],
        "title": "RuleDetail"
      },
...
    }
  }
}